remove k8s log extraction, add log structure abstraction
